get list timekeeping by date/month

+++ src/main/java/com/twendee/app/service/AdminService.java
@@ -10,7 +10,7 @@ public interface AdminService {
     public boolean refuseRequest(Integer requestId);
 
     public List<TimeKeepingDTO> getListTimekeepingByDate
-            (Integer page, Integer limit, Integer dateInt);
+            (Integer page, Integer limit, long dateInt);
 
     public List<TimeKeepingDTO> getListTimekeepingByMonth
             (Integer page, Integer limit, Integer month, Integer year);
